Rename operator_call_stack test=develop
