+++ paddle/fluid/framework/op_call_stack.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-#include "paddle/fluid/framework/operator_call_stack.h"
+#include "paddle/fluid/framework/op_call_stack.h"
 #include <string>
 #include <vector>
 #include "paddle/fluid/framework/attribute.h"
